feat: flex table width
