+++ cmd/tt/ui/pane.go
