+++ cmd/tt/ui/sidebarpane.go
