+++ cmd/tt/ui/sidebarpane_test.go
