+++ cmd/tt/ui/tasks_data.go
@@ -1,0 +1,512 @@
+package ui
+
+import "github.com/alswl/go-toodledo/pkg/models"
+
+func AllTasksMock() ([]*models.RichTask, error) {
+	return []*models.RichTask{
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+			TheGoal:    models.Goal{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "abc"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+		{
+			Task:       models.Task{Title: "def"},
+			TheContext: models.Context{},
+			TheFolder:  models.Folder{},
+		},
+	}, nil
+}
